requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
groq==0.4.2
gtts==2.4.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.27.2
pydub==0.25.1
deep-translator==1.11.4
